adding joplin_admin_search views
