+++ joplin/urls.py
@@ -9,10 +9,10 @@ from wagtail.admin import urls as wagtailadmin_urls
 from wagtail.core import urls as wagtail_urls
 from wagtail.documents import urls as wagtaildocs_urls
 from base.views import joplin_views
+from base.views import joplin_search_views
 from django.urls import reverse
 # from base.models import HomePage
 import debug_toolbar
-
 
 def home(request):
     # page = HomePage.objects.first()
@@ -35,7 +35,7 @@ urlpatterns = [
     url(r'admin/pages/(\d+)/publish/$', joplin_views.publish, name='publish'),
     url(r'admin/pages/new_from_modal/$',
         joplin_views.new_page_from_modal, name='new_page_from_modal'),
-    url(r'admin/pages/search/$', joplin_views.search, name='search' ),
+    url(r'admin/pages/search/$', joplin_search_views.search, name='search' ),
     url(r'^admin/', include(wagtailadmin_urls)),
     url(r'^documents/', include(wagtaildocs_urls)),
     path('__debug__/', include(debug_toolbar.urls)),
